requirements.txt
pytest>=7.2.0
selene>=2.0.0b16
allure-pytest
selenium